trading_system_backup_20250806_131207/requirements.txt
# External packages
numpy>=1.21.0
pandas>=1.3.0
ccxt>=4.0.0
aiohttp>=3.8.0
websockets>=10.0
requests>=2.25.0
python-dotenv>=0.19.0

# FastAPI and web
fastapi>=0.100.0
uvicorn>=0.20.0
jinja2>=3.1.0
websockets>=10.0

# Machine Learning
scikit-learn>=1.0.0
xgboost>=1.6.0
joblib>=1.1.0

# Technical Analysis
talib>=0.4.0
cachetools>=5.0.0

# OpenAI
openai>=1.0.0

# Notifications
feedparser>=6.0.0

# Data processing
pytz>=2021.3

# Security - API key encryption
cryptography>=41.0.0

# Optional dependencies for development
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0